+++ lib/ajax_datatables/filters.rb
@@ -1,7 +1,7 @@
 module AjaxDatatables
   class Filters
     include FormHelper  ## for ajax_draw(), ajax_search()
-    delegate :[], :each, :map, :reject, :flatten, :present?, :to_a, to: :@data
+    delegate :[], :each, :map, :select, :reject, :flatten, :present?, :to_a, :first, :last, to: :@data
     attr_accessor :datatable
 
     def initialize(ary = [], datatable: nil)