+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - spec: owg name correction
 - sorting_sql: super
 - spec: update details
+- ajax helper: find by xpath
 
 ### new feature